AbiTracer Program
This is the AbiTracer Program to read not only the header information as
the current Biopython package does, but also read the actual trace data
with peak heights for each of the four bases. This is important in the
evaluation of heterozygous mutations. This program uses an interface
based on the current .abi reader in Biopython by Wibowo Arindrarto, but
the heart of the actual .abi file reading is adapted from the
seqtrace-0.9.0 program by Brian J. Stucky.